+++ x/foundation/client/proposal_handler.go
@@ -1,0 +1,9 @@
+package client
+
+import (
+	"github.com/line/lbm-sdk/x/foundation/client/cli"
+	govclient "github.com/line/lbm-sdk/x/gov/client"
+)
+
+// ProposalHandler is the param change proposal handler.
+var ProposalHandler = govclient.NewProposalHandler(cli.NewProposalCmdFoundationExec)